requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
pydantic==2.8.2
SQLAlchemy==2.0.31
alembic==1.13.2
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.137.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
openai>=1.40.0
APScheduler==3.10.4
email-validator==2.2.0
stripe==8.11.0
cryptography==42.0.8
html2text==2024.2.26
pdfminer.six==20231228
python-dateutil==2.9.0.post0
pytz==2024.1
beautifulsoup4==4.12.3
lxml==5.2.2
dateparser==1.2.0
